refactor: consolidate shared JIVE layout and component helpers into JiveUtils
- Add source/UI/jive/JiveUtils.h with shared teardown and lookup helpers
- Replace duplicated clearJiveStyleSheets, collectJiveComponents, safeCleanupJiveTree, findComponentById, and findNodeById across 8 source/test files
- Add assertion and logging for WavExportTask non-modal loop fallback
- Register JiveUtils.h in CMake target sources
