+++ source/UI/jive/JiveModalDialog.cpp
@@ -1,45 +1,12 @@
 #include "UI/jive/JiveModalDialog.h"
 
 #include "UI/jive/DesignTokens.h"
+#include "UI/jive/JiveUtils.h"
 #include "UI/jive/StyleCatalog.h"
 
 namespace devpiano::ui::jive {
 
 namespace {
-
-// ============================================================================
-// JIVE Component Tree Safety Teardown Helpers
-// ============================================================================
-
-void clearJiveStyleSheets(juce::Component* comp) {
-    if (comp == nullptr) {
-        return;
-    }
-    for (int i = 0; i < comp->getNumChildComponents(); ++i) {
-        clearJiveStyleSheets(comp->getChildComponent(i));
-    }
-    if (comp->getProperties().contains("style-sheet")) {
-        comp->getProperties().remove("style-sheet");
-    }
-}
-
-void collectJiveComponents(::jive::GuiItem& item, std::vector<std::shared_ptr<juce::Component>>& components) {
-    if (auto component = item.getComponent()) {
-        components.push_back(std::move(component));
-    }
-    for (auto* child : item.getChildren()) {
-        collectJiveComponents(*child, components);
-    }
-}
-
-void safeCleanupJiveTree(std::unique_ptr<::jive::GuiItem>& rootItem) {
-    if (rootItem != nullptr) {
-        std::vector<std::shared_ptr<juce::Component>> jiveComponents;
-        collectJiveComponents(*rootItem, jiveComponents);
-        clearJiveStyleSheets(rootItem->getComponent().get());
-        rootItem.reset();
-    }
-}
 
 // ============================================================================
 // ValueTree Node Helpers
@@ -264,42 +231,6 @@ private:
 // ============================================================================
 // Public JiveModalDialog Methods
 // ============================================================================
-
-::jive::GuiItem* JiveModalDialog::findGuiItemById(::jive::GuiItem& root, const juce::String& id) {
-    if (root.state.getProperty("id").toString() == id) {
-        return &root;
-    }
-    for (auto* child : root.getChildren()) {
-        if (auto* found = findGuiItemById(*child, id)) {
-            return found;
-        }
-    }
-    return nullptr;
-}
-
-juce::Button* JiveModalDialog::findButtonById(::jive::GuiItem& root, const juce::String& id) {
-    if (root.state.getProperty("id").toString() == id) {
-        return dynamic_cast<juce::Button*>(root.getComponent().get());
-    }
-    for (auto* child : root.getChildren()) {
-        if (auto* b = findButtonById(*child, id)) {
-            return b;
-        }
-    }
-    return nullptr;
-}
-
-juce::TextEditor* JiveModalDialog::findTextEditorById(::jive::GuiItem& root, const juce::String& id) {
-    if (root.state.getProperty("id").toString() == id) {
-        return dynamic_cast<juce::TextEditor*>(root.getComponent().get());
-    }
-    for (auto* child : root.getChildren()) {
-        if (auto* ed = findTextEditorById(*child, id)) {
-            return ed;
-        }
-    }
-    return nullptr;
-}
 
 void JiveModalDialog::launchCustom(const LaunchOptions& options) {
     juce::DialogWindow::LaunchOptions opts;
